Implemented AddSupplierOrigin and Add Target Market - Also refactored Configurator to mulitple views
