+++ capgeminiappws2025/webapp/view/configurator/ConfiguratorTargetMarket.view.xml
@@ -1,0 +1,63 @@
+<mvc:View
+    controllerName="capgeminiappws2025.controller.configurator.ConfiguratorTargetMarket"
+    xmlns:mvc="sap.ui.core.mvc"
+    xmlns="sap.m"
+    xmlns:f="sap.ui.layout.form"
+>
+    <!-- ===== Target Market ===== -->
+    <HBox
+        justifyContent="SpaceBetween"
+        class="sapUiMediumMarginTop sapUiSmallMarginBegin sapUiSmallMarginEnd"
+    >
+        <Title
+            text="Target Market"
+            level="H6"
+        />
+        <Button
+            text="Add Target Market"
+            icon="sap-icon://add"
+            type="Transparent"
+            press=".onPressAddTargetMarket"
+        />
+    </HBox>
+
+    <Table
+        id="targetMarketTable"
+        class="sapUiMediumMargin"
+        width="90%"
+    >
+        <columns>
+            <Column>
+                <Text text="Country Code" />
+            </Column>
+            <Column
+                styleClass="sapUiSmallMarginBegin"
+                hAlign="End"
+            >
+                <Text text="Actions" />
+            </Column>
+        </columns>
+
+        <items>
+            <ColumnListItem>
+                <cells>
+                    <!-- Country Code: display only -->
+                    <Text text="{Country_Code}" />
+
+                    <!-- Actions: delete only (no edit) -->
+                    <HBox
+                        justifyContent="End"
+                        renderType="Bare"
+                    >
+                        <Button
+                            icon="sap-icon://delete"
+                            type="Reject"
+                            press=".onDeleteTargetMarket"
+                            tooltip="Delete target market"
+                        />
+                    </HBox>
+                </cells>
+            </ColumnListItem>
+        </items>
+    </Table>
+</mvc:View>
